src/cloud_function/requirements.txt
protobuf== 3.19.5
torch==1.10.1
google-cloud-storage